+++ app/controllers/web/admin/application_controller.rb
@@ -1,15 +1,12 @@
 # frozen_string_literal: true
 
 class Web::Admin::ApplicationController < Web::ApplicationController
-  before_action :authorize
+  before_action :require_admin
 
   private
 
-  def authorize(record = current_user, query = nil)
-    if record.present?
-      super([:admin, record], query)
-      return
-    end
+  def require_admin
+    return if current_user&.admin?
 
     flash[:error] = t('not_authorized')
     redirect_to request.referer || root_path
