+++ app/controllers/web/categories_controller.rb
@@ -3,7 +3,6 @@
 class Web::CategoriesController < Web::ApplicationController
   def show
     @category = Category.find(params[:id]).decorate
-    authorize @category
 
     scope = current_user&.admin? ? @category.bulletins : @category.bulletins.published
     @q = scope.ransack(params[:q])
