+++ app/controllers/web/users_controller.rb
@@ -4,9 +4,9 @@ class Web::UsersController < Web::ApplicationController
   before_action :authenticate_user
 
   def show
-    authorize(current_user)
+    @user = current_user
 
-    @q = current_user.bulletins.ransack(params[:q])
+    @q = @user.bulletins.ransack(params[:q])
     @q.sorts = 'updated_at asc' if @q.sorts.empty?
     @bulletins = @q.result(distinct: true).page(params[:page]).includes(image_attachment: :blob)
   end
