+++ app/policies/application_policy.rb
@@ -17,7 +17,7 @@ class ApplicationPolicy
   end
 
   def create?
-    user.present?
+    true
   end
 
   def new?
