--- app/policies/category_policy.rb
@@ -1,7 +1,0 @@
-# frozen_string_literal: true
-
-class CategoryPolicy < ApplicationPolicy
-  def show?
-    true
-  end
-end
