--- app/policies/user_policy.rb
@@ -1,13 +1,0 @@
-# frozen_string_literal: true
-
-class UserPolicy < ApplicationPolicy
-  def show?
-    self?
-  end
-
-  private
-
-  def self?
-    record == user
-  end
-end
