+++ app/views/web/admin/bulletins/shared/_action_buttons.html.slim
@@ -2,6 +2,7 @@ div id="#{dom_id(bulletin, :buttons)}"
   - if signed_in?
     .btn-group role='group' aria-label='Action buttons'
       = link_to_bulletin_action bulletin, :to_moderate, class: 'btn btn-sm btn-primary' if bulletin.may_to_moderate? && !bulletin.published?
-      = link_to_bulletin_action bulletin, :publish, class: 'btn btn-sm btn-primary' if bulletin.may_publish?
-      = link_to_bulletin_action bulletin, :reject, class: 'btn btn-sm btn-warning' if bulletin.may_reject?
+      - if current_user.admin?
+        = link_to_bulletin_action bulletin, :publish, class: 'btn btn-sm btn-primary' if bulletin.may_publish?
+        = link_to_bulletin_action bulletin, :reject, class: 'btn btn-sm btn-warning' if bulletin.may_reject?
       = link_to_bulletin_action bulletin, :archive, class: 'btn btn-sm btn-danger' if bulletin.may_archive?
